refactor: 📦 Enhance error handling in Downloader and Extractor
- Added error handling for HTTP errors, timeouts, and other standard errors.
- Provided custom error messages to give more context about the occurred error.
- Introduced a check to verify if the source gzipped file exists.
- Added error handling for invalid or corrupted gzip files and IO errors.
- Included custom error messages to offer a clearer understanding of the issues.

+++ lib/tasks/helpers/extractor.rb
@@ -2,6 +2,8 @@ require 'zlib'
 
 class Extractor
   def self.extract(gz_file_path, destination, records_amount)
+    raise "Source file #{gz_file_path} does not exist" unless File.exist?(gz_file_path)
+
     Zlib::GzipReader.open(gz_file_path) do |gz|
       File.open(destination, "wb") do |file|
         records_amount.times do
@@ -12,5 +14,9 @@ class Extractor
         end
       end
     end
+  rescue Zlib::GzipFile::Error
+    raise "The file #{gz_file_path} is not a valid gzip file or is corrupted"
+  rescue IOError => e
+    raise "An IO error occurred: #{e.message}"
   end
 end
